requirements.txt
colorama==0.4.6
Flask==2.3.2
Flask-SocketIO==5.3.4
importlib-metadata==6.8.0
numpy==1.24.3
google-generativeai==0.8.3
sentence-transformers==2.2.2
regex==2023.6.3
requests==2.31.0
tenacity==8.2.2
virtualenv==20.23.0
Werkzeug==3.0.3
Markdown==3.4.4
Pillow==10.3.0
Wikipedia-API==0.6.0
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
pyyaml==6.0
easydict==1.10
